move: #4 ディレクトリ名変更
