+++ rye-template/src/routers/health.py
