+++ rye-template/src/variables.py
